Rename Panel and UserPanel to Presence and UserPresence respectivly
