+++ .gitignore
@@ -332,4 +332,5 @@ cache
 top.db
 
 *.db
-
+*.db-shm
+*.db-wal
