+++ Cringe/Cringe.csproj
@@ -5,27 +5,27 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="BouncyCastle.NetCore" Version="1.8.8"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="5.0.6"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.6"/>
+        <PackageReference Include="BouncyCastle.NetCore" Version="1.8.8" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="5.0.6" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.6" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.0.6">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.2"/>
-        <PackageReference Include="OppaiSharp" Version="3.4.0"/>
+        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.2" />
+        <PackageReference Include="OppaiSharp" Version="3.4.0" />
     </ItemGroup>
 
     <ItemGroup>
-        <Folder Include="beatmapcache"/>
-        <Folder Include="wwwroot\images\"/>
+        <Folder Include="beatmapcache" />
+        <Folder Include="wwwroot\images\" />
     </ItemGroup>
 
     <ItemGroup>
-        <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap.min.css.map"/>
-        <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.min.js"/>
-        <None Include="wwwroot\lib\jquery-validation\dist\jquery.validate.min.js"/>
-        <None Include="wwwroot\lib\jquery\dist\jquery.min.map"/>
+        <None Include="wwwroot\lib\bootstrap\dist\css\bootstrap.min.css.map" />
+        <None Include="wwwroot\lib\bootstrap\dist\js\bootstrap.min.js" />
+        <None Include="wwwroot\lib\jquery-validation\dist\jquery.validate.min.js" />
+        <None Include="wwwroot\lib\jquery\dist\jquery.min.map" />
     </ItemGroup>
 
 </Project>
